Split convert and converter_registry to multiple smaller files
