+++ src/datumaro/v2/converters/__init__.py
@@ -2,5 +2,89 @@
 #
 # SPDX-License-Identifier: MIT
 
-from .converter_registry import ConverterRegistry, converter, find_conversion_path
-from .converters import *  # noqa: F403
+from datumaro.v2.converters.annotation_converters import (
+    LabelIndexConverter,
+    BBoxCoordinateConverter,
+    BBoxDtypeConverter,
+    LabelDtypeConverter,
+    PolygonToBBoxConverter,
+    RotatedBBoxToPolygonConverter,
+)
+from datumaro.v2.converters.image_converters import (
+    RGBToBGRConverter,
+    UInt8ToFloat32Converter,
+    ImagePathToImageConverter,
+    ImageToImageInfo,
+    ImageBytesToImageConverter,
+    ImageCallableToImageConverter,
+)
+from datumaro.v2.converters.mask_converters import (
+    PolygonToMaskConverter,
+    PolygonToInstanceMaskConverter,
+    InstanceMaskCallableToInstanceMaskConverter,
+    MaskCallableToMaskConverter,
+)
+from datumaro.v2.converters.registry import (
+    ConverterRegistry,
+    ConversionPaths,
+    converter,
+    _heuristic_cost,
+    _get_applicable_converters,
+    _group_fields_by_semantic,
+    _create_initial_renaming_converter,
+    _can_lazy_converter_handle_conversion,
+    _create_conversion_error_message,
+    _find_conversion_path_for_semantic,
+    _is_converter_lazy,
+    _separate_batch_and_lazy_converters,
+)
+from datumaro.v2.converters.base import (
+    Converter,
+    ConversionError,
+    AttributeRemapperConverter,
+    find_conversion_path,
+    ConverterTransform,
+    list_eval_ref,
+)
+
+__all__ = [
+    # Annotation converters
+    "LabelIndexConverter",
+    "BBoxCoordinateConverter",
+    "BBoxDtypeConverter",
+    "LabelDtypeConverter",
+    "PolygonToBBoxConverter",
+    "RotatedBBoxToPolygonConverter",
+    # Image converters
+    "RGBToBGRConverter",
+    "UInt8ToFloat32Converter",
+    "ImagePathToImageConverter",
+    "ImageToImageInfo",
+    "ImageBytesToImageConverter",
+    "ImageCallableToImageConverter",
+    # Mask converters
+    "PolygonToMaskConverter",
+    "PolygonToInstanceMaskConverter",
+    "InstanceMaskCallableToInstanceMaskConverter",
+    "MaskCallableToMaskConverter",
+    # Registry
+    "ConverterRegistry",
+    "ConversionPaths",
+    "converter",
+    "_heuristic_cost",
+    "_get_applicable_converters",
+    "_group_fields_by_semantic",
+    "_create_initial_renaming_converter",
+    "_can_lazy_converter_handle_conversion",
+    "_create_conversion_error_message",
+    "_find_conversion_path_for_semantic",
+    "_is_converter_lazy",
+    "_separate_batch_and_lazy_converters",
+    # Base
+    "Converter",
+    "ConversionError",
+    "AttributeRemapperConverter",
+    "find_conversion_path",
+    "ConverterTransform",
+    "list_eval_ref",
+]
